file re-organization. index size trimming.
