+++ src/commands/mock.ts
@@ -47,6 +47,7 @@ export const Mock = {
         }
         mockTargets.add(target.id);
         await interaction.reply(`Sure thing, I will begin mocking ${target}!`);
+        console.log(`Current Mock list: ${mockTargets}`);
     },
     /**
      * Deletes 'mock' targets message and re-sends it in "sPoNgEbOb" text format
@@ -69,7 +70,6 @@ export const Mock = {
                 shouldBeLower = !shouldBeLower;
             }
         }
-
         // Attempt to delete the original message
         try {
             await message.delete();
@@ -78,7 +78,6 @@ export const Mock = {
             console.log(`ERROR: could not delete ${message.author.username} (${message.author.id}) message in ${message.channel}`);
             return;
         }
-
         // Attempt to send the mocked message
         try {
             await message.channel.send(`${message.author} says "${mockMessage}"`);
